requirements.txt
flask
line-bot-sdk
psycopg2
prettytable